Separate out fuel pump controller
